Started with Database API.
Most database functions will now work and output a formatted query, so they can
already be used for testing. They won't yet give a result though.

Note that, in this version, addslashes is used to escape strings in queries. This is
not correct and/or safe but prevents the need to connect to a database already.