Added userlist (not finished), other minor changes
Userlist is not working yet. I want to have the second option working, but somewhere is a mistake (somewhere in extjs a substring method fails). The first option doesn't work with paging, so if I can't fix the second option, then I'm going back to the first one and try to find the problem over there.
COLLAB-52