Added pagination to search results, changed input format
In order to add pagination to the search results, I had to change to input format from nice object oriented JSON to a array-oriented JSON (which is actually shorter in terms of transmitted bytes per search request and forces fields to be in the same order every time, which is a good thing).
Also, the refresh button for the pagination toolbar has been removed, as using it would not make sense (refreshing a memory store).

COLLAB-51