Finalize Proxy for RequestManager.
In Controller, translator has been commented out as it gives errors about redeclaring _(). Also, the JSON processing logic is fixed (it did not work at all because of escaping), but the new POST model has not yet been implemented.
Added UserController to serve a static user list with pagination.

COLLAB-65