New limitation to username, better check for email/username
Limitation to username is as agreed at TD meeting: a-zA-Z0-9 .'_@
actionUsernameExists also checks if there's already an email address with the same input. actionEmailExists also checks if there's already an username with the same input. It doesn't check with itself, so you can have the same email address and username.
COLLAB-253