Created the views for reordening scans and selecting books
The views only contain some basic elements, I hope enough to get Jeroen and Tom W. started.
I also added multiselect.js, as this was necessary for the reordening of scans. I also decided to decouple the two pages, which makes it a lot easier.
Also changed/added some minor things, like: all columns of userlist are now flex: 1, which is better for small/large screens and for zooming in/out.
Also added a bindingcontroller.php, but it doesn't do much yet. I don't think it's something to start with: redoing my work is a better option I assume.