Activation screen finally works.
Therefore user activation via e-mail is practically done. Note that some
settings need to be specified in the database (activation-mail-subject and
activation-mail-message) need to be specified for this to work correctly.
Additionally, the sendmail utility should be present, PHP should be able to
find its executable and it should be properly configured.

Also, perhaps the from-address configuration needs to be changed in order
to prevent e-mails from being blocked by spam filters.

COLLAB-147