This should add fulltext indices to the database.
Indices are made on fulltext columns mirrored in MyISAM tables, which are automatically updated by means of triggers.
This approach is however not transaction-safe, so in rare circumstances the fulltext index may be corrupted (spurious records may appear or records may be missing), so it might be a good idea to update the indices every once in a while.