Issue #37: bindings can now be edited by non-owners.
Whenever a moderator or binding creator wants to start modifying a binding
or whenever a user starts an upload, the currentBindingId column of that
user will be set to the binding that is being edited.

Furthermore, GetBindingStatus will now look for this binding instead of
providing a reordered binding the user owns.